Implement config file parsing, add strncmp
Add strncmp() to compare strings more easily, memcmp()
is not flexible enough.

Implement a config file parser using a FSM and gotos.
While gotos are said to be evil, they work pretty well
when implementing FSMs, such as language parsers, so
i have decided to use them.

The configuration syntax is very simple, but it should
work just fine for such simple case, all the project
needs as of now are 3 config options.