Copy code from bootloader to display ACPI in kernel
Add display of ACPI pointer in kernel to make sure it receives
it correctly. This is done by copying files from bootloader.

All the dependency of kernel on UEFI is very ugly and this
commit is probably the breaking point before a proper ExitBootServices()
call is added before passing control to the kernel. Most currently existing
kernel code will most likely be removed or rewritten in the future.